Make pag-screen more robust to missing screen
It will now attempt to find screen in your path.
Failing that, it will try to find attachandrun in your path.
If that succeeds, it will use attachandrun --check to see if screen is
available.

If screen does not seem to be available, it reports and error and exits
before doing anything else. If screen is available, it will run the version
in the PATH or via attachandrun of the gnu locker, in that order of
preference.