Allow programs to be missing in PATH
In 89582ba5b09ff1c72, we began using set -e. This means that checking for the
path to screen or cont-renew-notify became problematic --- if it appears in the
PATH, everything is fine, but if it is missing, `which` will fail and the
script will exit. By running which || :, we allow failures.